Add 100 Beers

+++ domain/src/main/java/com/pabji/domain/model/Beer.kt
@@ -6,7 +6,7 @@ data class Beer(
     val description: String = "",
     val imageUrl: String = "",
     val abv: Double = 0.0,
-    val ibu: Int = 0,
+    val ibu: Double = 0.0,
     val foodPairing: List<String> = emptyList(),
     val isBarrelEmpty: Boolean = false
 )
